Revert LoadClass moving
