+++ sentry-android-core/api/sentry-android-core.api
@@ -1,5 +1,5 @@
 public final class io/sentry/android/core/ActivityFramesTracker {
-	public fun <init> (Lio/sentry/util/LoadClass;)V
+	public fun <init> (Lio/sentry/android/core/LoadClass;)V
 	public fun addActivity (Landroid/app/Activity;)V
 	public fun setMetrics (Landroid/app/Activity;Lio/sentry/protocol/SentryId;)V
 	public fun stop ()V
@@ -78,6 +78,11 @@ public abstract interface class io/sentry/android/core/IBuildInfoProvider {
 public abstract interface class io/sentry/android/core/IDebugImagesLoader {
 	public abstract fun clearDebugImages ()V
 	public abstract fun loadDebugImages ()Ljava/util/List;
+}
+
+public final class io/sentry/android/core/LoadClass {
+	public fun <init> ()V
+	public fun loadClass (Ljava/lang/String;)Ljava/lang/Class;
 }
 
 public final class io/sentry/android/core/NdkIntegration : io/sentry/Integration, java/io/Closeable {
@@ -173,7 +178,7 @@ public final class io/sentry/android/core/TempSensorBreadcrumbsIntegration : and
 }
 
 public final class io/sentry/android/core/UserInteractionIntegration : android/app/Application$ActivityLifecycleCallbacks, io/sentry/Integration, java/io/Closeable {
-	public fun <init> (Landroid/app/Application;Lio/sentry/util/LoadClass;)V
+	public fun <init> (Landroid/app/Application;Lio/sentry/android/core/LoadClass;)V
 	public fun close ()V
 	public fun onActivityCreated (Landroid/app/Activity;Landroid/os/Bundle;)V
 	public fun onActivityDestroyed (Landroid/app/Activity;)V
