+++ sentry-android-core/src/main/java/io/sentry/android/core/LoadClass.java
@@ -1,4 +1,4 @@
-package io.sentry.util;
+package io.sentry.android.core;
 
 import org.jetbrains.annotations.NotNull;
