+++ sentry-android-core/src/main/java/io/sentry/android/core/UserInteractionIntegration.java
@@ -12,7 +12,6 @@ import io.sentry.SentryOptions;
 import io.sentry.android.core.internal.gestures.NoOpWindowCallback;
 import io.sentry.android.core.internal.gestures.SentryGestureListener;
 import io.sentry.android.core.internal.gestures.SentryWindowCallback;
-import io.sentry.util.LoadClass;
 import io.sentry.util.Objects;
 import java.io.Closeable;
 import java.io.IOException;
